requirements.txt
gradio==4.44.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
diffusers==0.24.0
transformers==4.36.0
accelerate==0.25.0
opencv-python==4.8.1.78
Pillow==10.1.0
moviepy==1.0.3
requests==2.31.0
openai-whisper==20231117
google-generativeai==0.3.2
imageio[ffmpeg]==2.33.1
scikit-learn==1.3.2
numpy==1.24.3
